\name{cRacker-package}
\alias{cRacker-package}
\alias{cRacker}
\docType{package}
\title{
a R-script suite for automated processing of proteomic peptide lists
}
\description{
cRacker has been programmed to automate standard normalization strategies of quantitative proteomic experiments. It simplifies the analysis of quantified peptide lists coming from various proteomic quantitation softwares. The applied normalization strategies are optimized for label free analyses of peptide intensities but also experiments using labeling approaches can be analyzed. 
}
\details{
\tabular{ll}{
Package: \tab cRacker\cr
Type: \tab Package\cr
Version: \tab 1.488\cr
Date: \tab 2012-09-11\cr
License: \tab  GNU General Public License\cr
}


}
\author{
Henrik Zauber

Maintainer: Henrik Zauber <Zauber@mpimp-golm.mpg.de>
}
\references{
}

\keyword{proteomics}
\keyword{peptide}
\keyword{protein}
\keyword{intensity}
\keyword{emPAI}
% \seealso{
% ~~ Optional links to other man pages, e.g. ~~
% ~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
% }
\examples{
cRacker.starter()
}
